Completing server microservices
Renamed file system service to storage
Added client package to storage service
Dependencies have been tidied up
