+++ go.mod
@@ -1,0 +1,33 @@
+module remote-storage
+
+go 1.20
+
+require (
+	github.com/go-kit/kit v0.12.0
+	github.com/golang-jwt/jwt/v4 v4.5.0
+	github.com/golang-migrate/migrate/v4 v4.16.2
+	github.com/gorilla/mux v1.8.0
+	github.com/hashicorp/consul/api v1.10.1
+	github.com/lib/pq v1.10.9
+)
+
+require (
+	github.com/armon/go-metrics v0.3.9 // indirect
+	github.com/fatih/color v1.12.0 // indirect
+	github.com/go-kit/log v0.2.0 // indirect
+	github.com/go-logfmt/logfmt v0.5.1 // indirect
+	github.com/hashicorp/errwrap v1.1.0 // indirect
+	github.com/hashicorp/go-cleanhttp v0.5.2 // indirect
+	github.com/hashicorp/go-hclog v0.16.2 // indirect
+	github.com/hashicorp/go-immutable-radix v1.3.1 // indirect
+	github.com/hashicorp/go-multierror v1.1.1 // indirect
+	github.com/hashicorp/go-rootcerts v1.0.2 // indirect
+	github.com/hashicorp/golang-lru v0.5.4 // indirect
+	github.com/hashicorp/serf v0.9.5 // indirect
+	github.com/mattn/go-colorable v0.1.13 // indirect
+	github.com/mattn/go-isatty v0.0.19 // indirect
+	github.com/mitchellh/go-homedir v1.1.0 // indirect
+	github.com/mitchellh/mapstructure v1.4.2 // indirect
+	go.uber.org/atomic v1.9.0 // indirect
+	golang.org/x/sys v0.10.0 // indirect
+)
