+++ server/.idea/workspace.xml
@@ -5,14 +5,11 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="454c6521-c7f9-4f7b-a23b-1cfabc86092f" name="Changes" comment="">
+      <change beforePath="$PROJECT_DIR$/../.gitignore" beforeDir="false" afterPath="$PROJECT_DIR$/../.gitignore" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/file_system_svc/endpoints.go" beforeDir="false" afterPath="$PROJECT_DIR$/file_system_svc/endpoints.go" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/auth_svc/test.http" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/file_system_svc/config/config.json" beforeDir="false" afterPath="$PROJECT_DIR$/file_system_svc/config/config.json" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/file_system_svc/log/log.txt" beforeDir="false" afterPath="$PROJECT_DIR$/file_system_svc/log/log.txt" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/file_system_svc/logging_middleware.go" beforeDir="false" afterPath="$PROJECT_DIR$/file_system_svc/logging_middleware.go" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/file_system_svc/repository/filesystem/commands.go" beforeDir="false" afterPath="$PROJECT_DIR$/file_system_svc/repository/filesystem/commands.go" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/file_system_svc/service.go" beforeDir="false" afterPath="$PROJECT_DIR$/file_system_svc/service.go" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/file_system_svc/testsvc.http" beforeDir="false" afterPath="$PROJECT_DIR$/file_system_svc/testsvc.http" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/file_system_svc/transport.go" beforeDir="false" afterPath="$PROJECT_DIR$/file_system_svc/transport.go" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -50,28 +47,30 @@
     "WebServerToolWindowFactoryState": "false",
     "go.import.settings.migrated": "true",
     "go.sdk.automatically.set": "true",
-    "last_opened_file_path": "D:/Go/remote-storage/server",
+    "last_opened_file_path": "D:/Go/remote-storage",
     "node.js.detected.package.eslint": "true",
     "node.js.selected.package.eslint": "(autodetect)"
   }
 }]]></component>
   <component name="RecentsManager">
     <key name="CopyFile.RECENT_KEYS">
+      <recent name="D:\Go\remote-storage\server\authsvc" />
       <recent name="D:\Go\remote-storage\server\common" />
       <recent name="D:\Go\remote-storage\server\auth_svc" />
     </key>
     <key name="MoveFile.RECENT_KEYS">
+      <recent name="D:\Go\remote-storage\server" />
       <recent name="D:\Go\remote-storage\server\auth_svc" />
       <recent name="D:\Go\remote-storage\server\file_system_service" />
       <recent name="D:\Go\remote-storage\server\file_system_service\internal" />
     </key>
   </component>
   <component name="RunManager" selected="Go Build.go build server/file_system_svc/cmd/app">
-    <configuration name="go build server/auth_svc/cmd/app" type="GoApplicationRunConfiguration" factoryName="Go Application" nameIsGenerated="true">
+    <configuration name="go build server/authsvc/cmd/app" type="GoApplicationRunConfiguration" factoryName="Go Application">
       <module name="server" />
       <working_directory value="$PROJECT_DIR$" />
       <kind value="PACKAGE" />
-      <package value="server/auth_svc/cmd/app" />
+      <package value="server/authsvc/cmd/app" />
       <directory value="$PROJECT_DIR$" />
       <filePath value="$PROJECT_DIR$" />
       <method v="2" />
@@ -88,10 +87,10 @@
     <configuration name="All in testsvc" type="HttpClient.HttpRequestRunConfigurationType" factoryName="HTTP Request" temporary="true" path="$PROJECT_DIR$/file_system_svc/testsvc.http">
       <method v="2" />
     </configuration>
-    <configuration name="test | #1" type="HttpClient.HttpRequestRunConfigurationType" factoryName="HTTP Request" temporary="true" nameIsGenerated="true" path="$PROJECT_DIR$/auth_svc/test.http" requestIdentifier="#1" runType="Run single request">
+    <configuration name="test | #1 (1)" type="HttpClient.HttpRequestRunConfigurationType" factoryName="HTTP Request" temporary="true" nameIsGenerated="true" path="$PROJECT_DIR$/authsvc/test.http" requestIdentifier="#1" runType="Run single request">
       <method v="2" />
     </configuration>
-    <configuration name="test | #2" type="HttpClient.HttpRequestRunConfigurationType" factoryName="HTTP Request" temporary="true" nameIsGenerated="true" path="$PROJECT_DIR$/auth_svc/test.http" index="2" requestIdentifier="#2" runType="Run single request">
+    <configuration name="test | #2" type="HttpClient.HttpRequestRunConfigurationType" factoryName="HTTP Request" temporary="true" nameIsGenerated="true" path="$PROJECT_DIR$/authsvc/test.http" index="2" requestIdentifier="#2" runType="Run single request">
       <method v="2" />
     </configuration>
     <configuration name="testsvc | #1 (2)" type="HttpClient.HttpRequestRunConfigurationType" factoryName="HTTP Request" temporary="true" nameIsGenerated="true" path="$PROJECT_DIR$/file_system_svc/testsvc.http" requestIdentifier="#1" runType="Run single request">
@@ -101,21 +100,21 @@
       <method v="2" />
     </configuration>
     <list>
+      <item itemvalue="Go Build.go build server/authsvc/cmd/app" />
       <item itemvalue="Go Build.go build server/file_system_svc/cmd/app" />
-      <item itemvalue="Go Build.go build server/auth_svc/cmd/app" />
+      <item itemvalue="HTTP Request.test | #2" />
+      <item itemvalue="HTTP Request.test | #1 (1)" />
       <item itemvalue="HTTP Request.All in testsvc" />
       <item itemvalue="HTTP Request.testsvc | #2" />
-      <item itemvalue="HTTP Request.test | #2" />
       <item itemvalue="HTTP Request.testsvc | #1 (2)" />
-      <item itemvalue="HTTP Request.test | #1" />
     </list>
     <recent_temporary>
       <list>
-        <item itemvalue="HTTP Request.testsvc | #1 (2)" />
         <item itemvalue="HTTP Request.testsvc | #2" />
-        <item itemvalue="HTTP Request.All in testsvc" />
-        <item itemvalue="HTTP Request.test | #1" />
         <item itemvalue="HTTP Request.test | #2" />
+        <item itemvalue="HTTP Request.test | #1 (1)" />
+        <item itemvalue="HTTP Request.testsvc | #1 (2)" />
+        <item itemvalue="HTTP Request.All in testsvc" />
       </list>
     </recent_temporary>
   </component>
