+++ server/authsvc/test.http
@@ -1,0 +1,22 @@
+POST http://localhost:666/authentication/login
+Content-Type: application/json
+
+{
+  "login": "test",
+  "hashed_password": "qwerty"
+}
+
+###
+POST http://localhost:666/authentication/validate
+Content-Type: application/json
+
+{
+  "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6InRlc3QiLCJyb290X2RpciI6InRlc3QiLCJleHAiOjE2OTA5MTkyMjZ9.HtRfccRL9MMmsx1qq1LunB06Qs35o9WtKvyHr5RlqPw"
+}
+
+###
+
+GET http://localhost:666/authentication/test
+Accept: application/json
+
+###
