--- server/file_system_svc/config/config.json
@@ -1,3 +1,0 @@
-{
-  "RootDirectory": "D:\\Go\\remote-storage\\server\\storage\\"
-}
