+++ server/storagesvc/testsvc.http
@@ -1,0 +1,24 @@
+
+POST http://localhost:777/filesystem/download
+Cookie: token=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6InRlc3QiLCJyb290X2RpciI6InRlc3QiLCJleHAiOjE2OTExNDcyNjJ9.alMB8UbnKCfH36h_HkB67VZ1PKaH_yIa8bPIJYayq1s
+Content-Type: application/json
+
+{
+  "dir_path": "api/",
+  "file_name": "server.go"
+}
+
+
+###
+
+POST http://localhost:777/filesystem/rename
+Cookie: token=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6InRlc3QiLCJyb290X2RpciI6InRlc3QiLCJleHAiOjE2OTExNDcyNjJ9.alMB8UbnKCfH36h_HkB67VZ1PKaH_yIa8bPIJYayq1s
+Content-Type: application/json
+
+{
+  "dir_path": "api/",
+  "old_name": "not_server.go",
+  "new_name": "server.go"
+}
+
+
